// web/src/generated/models/NotificationSourcesUsableResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Revornix Main Backend
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.6.4
 * Contact: 1142704468@qq.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { NotificationSource } from './NotificationSource';
import {
    NotificationSourceFromJSON,
    NotificationSourceFromJSONTyped,
    NotificationSourceToJSON,
    NotificationSourceToJSONTyped,
} from './NotificationSource';

/**
 * 
 * @export
 * @interface NotificationSourcesUsableResponse
 */
export interface NotificationSourcesUsableResponse {
    /**
     * 
     * @type {Array<NotificationSource>}
     * @memberof NotificationSourcesUsableResponse
     */
    data: Array<NotificationSource>;
}

/**
 * Check if a given object implements the NotificationSourcesUsableResponse interface.
 */
export function instanceOfNotificationSourcesUsableResponse(value: object): value is NotificationSourcesUsableResponse {
    if (!('data' in value) || value['data'] === undefined) return false;
    return true;
}

export function NotificationSourcesUsableResponseFromJSON(json: any): NotificationSourcesUsableResponse {
    return NotificationSourcesUsableResponseFromJSONTyped(json, false);
}

export function NotificationSourcesUsableResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): NotificationSourcesUsableResponse {
    if (json == null) {
        return json;
    }
    return {
        
        'data': ((json['data'] as Array<any>).map(NotificationSourceFromJSON)),
    };
}

export function NotificationSourcesUsableResponseToJSON(json: any): NotificationSourcesUsableResponse {
    return NotificationSourcesUsableResponseToJSONTyped(json, false);
}

export function NotificationSourcesUsableResponseToJSONTyped(value?: NotificationSourcesUsableResponse | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'data': ((value['data'] as Array<any>).map(NotificationSourceToJSON)),
    };
}