// web/src/generated/models/NotificationTargetsUsableResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Revornix Main Backend
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.6.4
 * Contact: 1142704468@qq.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { NotificationTarget } from './NotificationTarget';
import {
    NotificationTargetFromJSON,
    NotificationTargetFromJSONTyped,
    NotificationTargetToJSON,
    NotificationTargetToJSONTyped,
} from './NotificationTarget';

/**
 * 
 * @export
 * @interface NotificationTargetsUsableResponse
 */
export interface NotificationTargetsUsableResponse {
    /**
     * 
     * @type {Array<NotificationTarget>}
     * @memberof NotificationTargetsUsableResponse
     */
    data: Array<NotificationTarget>;
}

/**
 * Check if a given object implements the NotificationTargetsUsableResponse interface.
 */
export function instanceOfNotificationTargetsUsableResponse(value: object): value is NotificationTargetsUsableResponse {
    if (!('data' in value) || value['data'] === undefined) return false;
    return true;
}

export function NotificationTargetsUsableResponseFromJSON(json: any): NotificationTargetsUsableResponse {
    return NotificationTargetsUsableResponseFromJSONTyped(json, false);
}

export function NotificationTargetsUsableResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): NotificationTargetsUsableResponse {
    if (json == null) {
        return json;
    }
    return {
        
        'data': ((json['data'] as Array<any>).map(NotificationTargetFromJSON)),
    };
}

export function NotificationTargetsUsableResponseToJSON(json: any): NotificationTargetsUsableResponse {
    return NotificationTargetsUsableResponseToJSONTyped(json, false);
}

export function NotificationTargetsUsableResponseToJSONTyped(value?: NotificationTargetsUsableResponse | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'data': ((value['data'] as Array<any>).map(NotificationTargetToJSON)),
    };
}