// web/src/generated/models/PaginationNotificationTask.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Revornix Main Backend
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.5.1
 * Contact: 1142704468@qq.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { NotificationTask } from './NotificationTask';
import {
    NotificationTaskFromJSON,
    NotificationTaskFromJSONTyped,
    NotificationTaskToJSON,
    NotificationTaskToJSONTyped,
} from './NotificationTask';

/**
 * 
 * @export
 * @interface PaginationNotificationTask
 */
export interface PaginationNotificationTask {
    /**
     * 
     * @type {number}
     * @memberof PaginationNotificationTask
     */
    total_elements: number;
    /**
     * 
     * @type {number}
     * @memberof PaginationNotificationTask
     */
    current_page_elements: number;
    /**
     * 
     * @type {number}
     * @memberof PaginationNotificationTask
     */
    total_pages: number;
    /**
     * 
     * @type {number}
     * @memberof PaginationNotificationTask
     */
    page_num: number;
    /**
     * 
     * @type {number}
     * @memberof PaginationNotificationTask
     */
    page_size: number;
    /**
     * 
     * @type {Array<NotificationTask>}
     * @memberof PaginationNotificationTask
     */
    elements: Array<NotificationTask>;
}

/**
 * Check if a given object implements the PaginationNotificationTask interface.
 */
export function instanceOfPaginationNotificationTask(value: object): value is PaginationNotificationTask {
    if (!('total_elements' in value) || value['total_elements'] === undefined) return false;
    if (!('current_page_elements' in value) || value['current_page_elements'] === undefined) return false;
    if (!('total_pages' in value) || value['total_pages'] === undefined) return false;
    if (!('page_num' in value) || value['page_num'] === undefined) return false;
    if (!('page_size' in value) || value['page_size'] === undefined) return false;
    if (!('elements' in value) || value['elements'] === undefined) return false;
    return true;
}

export function PaginationNotificationTaskFromJSON(json: any): PaginationNotificationTask {
    return PaginationNotificationTaskFromJSONTyped(json, false);
}

export function PaginationNotificationTaskFromJSONTyped(json: any, ignoreDiscriminator: boolean): PaginationNotificationTask {
    if (json == null) {
        return json;
    }
    return {
        
        'total_elements': json['total_elements'],
        'current_page_elements': json['current_page_elements'],
        'total_pages': json['total_pages'],
        'page_num': json['page_num'],
        'page_size': json['page_size'],
        'elements': ((json['elements'] as Array<any>).map(NotificationTaskFromJSON)),
    };
}

export function PaginationNotificationTaskToJSON(json: any): PaginationNotificationTask {
    return PaginationNotificationTaskToJSONTyped(json, false);
}

export function PaginationNotificationTaskToJSONTyped(value?: PaginationNotificationTask | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'total_elements': value['total_elements'],
        'current_page_elements': value['current_page_elements'],
        'total_pages': value['total_pages'],
        'page_num': value['page_num'],
        'page_size': value['page_size'],
        'elements': ((value['elements'] as Array<any>).map(NotificationTaskToJSON)),
    };
}